# Fix performance issues when creating multiple instances of Period (#12903, #11831)