--- a/asv_bench/benchmarks/period.py
+++ b/asv_bench/benchmarks/period.py
@@ -1,4 +1,4 @@
-from pandas import Period, PeriodIndex, date_range
+from pandas import PeriodIndex, date_range
 
 
 class create_period_index_from_date_range(object):
@@ -6,4 +6,4 @@ class create_period_index_from_date_range(object):
 
     def time_period_index(self):
         # Simulate irregular PeriodIndex
-        PeriodIndex(date_range('1985', periods=10000).to_pydatetime(), freq='D')
+        PeriodIndex(date_range('1985', periods=1000).to_pydatetime(), freq='D')