--- a/pandas/tseries/constants.py
+++ b/pandas/tseries/constants.py
@@ -1,0 +1,7 @@
+US_RESO = 0
+MS_RESO = 1
+S_RESO = 2
+T_RESO = 3
+H_RESO = 4
+D_RESO = 5
+