@jreback 2016-04-16
pls run the asv suite for relevant benchmarks