@jreback 2016-04-16
needs a whatsnew entry