@jreback 2016-04-17
also pls show the asv runs for the period benchmarks