@rs2 2016-04-20
@jreback , @tswicegood Is there a doc describing the hoops I need to jump through to get Travis and Jeff happy to merge? :smile: I.e. the names of the scripts I need to run to add asv benchmarks, pass coverage and pep8 requirements?