@jorisvandenbossche 2016-04-20
@rs2 see @jreback's previous comments:

- add a note in the whatsnew file about the performance improvement (doc/source/whatsnew/v0.18.1)
- run the asv benchmarks and paste the results here, see http://pandas-docs.github.io/pandas-docs-travis/contributing.html#running-the-performance-test-suite (you can select the relevant ones to run, related to periods and timeseries plotting)
- for travis, you just have a failing test for which you need to fix the code, it seems related to the moving of the constants.