@jreback 2016-04-25
can you show some benchmark results (prob need some asv's for this as well).