@jreback 2016-04-26
just force push to this PR
opening multiple PRs for the same issue is confusing and more work for others