@rs2 2016-04-26
The [failing test](https://travis-ci.org/pydata/pandas/jobs/125805786) is not mine.