@rs2 2016-04-26
Adding `asv` benchmarks...