@rs2 2016-04-26
### `asv` benchmarks

From 5.89s to 525ms:

#### Before the change:
```
~/pandas_master/asv_bench$ asv dev -b create_period
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Building for /home/pashok/anaconda3/bin/python
[  0.00%] ·· Benchmarking /home/pashok/anaconda3/bin/python
[100.00%] ··· Running period.create_period_index_from_date_range.time_period_index                                                                                                    5.89s
```

#### After the change
```
~/pandas/asv_bench$ asv dev -b create_period
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Building for /home/pashok/anaconda3/bin/python
[  0.00%] ·· Benchmarking /home/pashok/anaconda3/bin/python
[100.00%] ··· Running period.create_period_index_from_date_range.time_period_index                                                                                   525.84ms
```