@jorisvandenbossche 2016-04-26
Nice speedup! Did you run the benchmarks on py2 or py3?