@jreback 2016-04-26
``~/pandas/asv_bench$ asv continuous -b period`` (``-b`` just matches so you can see results (will compare)