@rs2 2016-04-26
@jorisvandenbossche : plot `asv` benchmarks:

#### Before the change
```
~/pandas_master/asv_bench$ asv dev -b foo
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Building for /home/pashok/anaconda3/bin/python
[  0.00%] ·· Benchmarking /home/pashok/anaconda3/bin/python
[ 50.00%] ··· Running foo_bar.test_foo_bar.time_plot_date_ranged_df                                                                                                                   1.81s
[100.00%] ··· Running foo_bar.test_foo_bar.time_plot_plain_df                                                                                                                      129.33ms
~/pandas_master/asv_bench$ python --version
Python 3.5.1 :: Anaconda 2.5.0 (64-bit)
pashok@pashok-VirtualBox:~/pandas_master/asv_bench$
```

#### After the change
```
~/pandas/asv_bench$ asv dev -b foo
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Building for /home/pashok/anaconda3/bin/python
[  0.00%] ·· Benchmarking /home/pashok/anaconda3/bin/python
[ 50.00%] ··· Running foo_bar.test_foo_bar.time_plot_date_ranged_df                                                                                                  214.24ms
[100.00%] ··· Running foo_bar.test_foo_bar.time_plot_plain_df                                                                                                        123.06ms
~/pandas/asv_bench$ python --version
Python 3.5.1 :: Anaconda 2.5.0 (64-bit)
~/pandas/asv_bench$
```

```python
import numpy as np
import pandas as pd


class test_foo_bar(object):

    goal_time = 0.2
    N = 2000

    def setup(self):
        self.date_ranged_df = pd.DataFrame(np.random.randn(self.N, 5), index=pd.date_range('1/1/1975', periods=self.N))
        self.plain_df = pd.DataFrame(np.random.randn(self.N, 5))

    def time_plot_date_ranged_df(self):
        self.date_ranged_df.plot()

    def time_plot_plain_df(self):
        self.plain_df.plot()
```