@blbradley 2016-04-26
This was a big TODO I had when moving Period to Cython. Thank you for working it out @rs2.